Move MMAP* to crate::mem
