+++ mizu/kernel/src/syscall.rs
@@ -28,9 +28,9 @@ pub static SYSCALL: Lazy<AHandlers<Scn, ScParams, ScRet>> = Lazy::new(|| {
         .map(FUTEX, crate::mem::futex)
         .map(GET_ROBUST_LIST, crate::mem::get_robust_list)
         .map(SET_ROBUST_LIST, crate::mem::set_robust_list)
-        .map(MMAP, fd::mmap)
-        .map(MPROTECT, fd::mprotect)
-        .map(MUNMAP, fd::munmap)
+        .map(MMAP, crate::mem::mmap)
+        .map(MPROTECT, crate::mem::mprotect)
+        .map(MUNMAP, crate::mem::munmap)
         // Tasks
         .map(SCHED_YIELD, task::uyield)
         .map(GETTID, task::tid)
